LPS-163083 Move management toolbar display ctx to its own file
